Temporarily remove section collation logic
a) Don't have a .ts test pattern that has a table with multiple
   sections
b) Need to do other refactoring and this area needs to be simpler to
   support that
c) Maybe multi-section tables aren't very common in real usage
   (unsubstantiated)

More things need refactoring to work in terms of borrowed data, and the
owned Vec here seems to complicate that.

Logic to collect together all the parts of a new table before supplying
the collated result to the next processing stage can be added back
later.