Don't carry TableProcessor in TableSectionConsumer
Rather, the (increasingly misnamed) TableSectionConsumer will return the
resulting TableSection object, which will be passed to
TableSectionConsumer by the calling code.

This removes some references to demultiplex::FilterChangeset within the
PSI code (but not yet all; must try harder).

This decoupling makes things a little but easier to understand, IMO.