Avoid 'unreachable pattern' warning
Force user of descriptor_enum!{} to cover all possible variants in their
definition.

Think this may make the code incompatible with older versions of rust,
which do not do exhaustiveness checks in integer matches.  Therefore
might revert this and look for another approach if this way turns out to
be problematic for some users.