Convert StreamType and StreamId types to structs
The use of enums for these types presented a problem for backwards
compatibility.  If code was written to make use of the Reserved(u8) or
Unknown(u8) variants, and then in future we add an explicit variant to the enum
of that particular value becomes defined by the spec, then code expecting to
expect that value to be available in e.g. Reserved will silently break - it
will not hit a compilation error, but it will not do the expected thing any
more at runtime.  By converting these types to simply u8 wrappers calling code
can just construct their own instances for values not currently defined in the
spec, and those values will be compatible with any definitions that this crate
adds in future.

Benchmarking shows a very slight performance improvement with this change (sub
1%).